[Refactor] Post 도메인 관련 리팩토링 및 버그 수정 다수 (#336)
* Feat: 댓글 등록 기능 구현

* Refactor: 일단 PickApplication이 돌아는 가게 수정

* Refactor: 일단 PickApplication이 돌아는 가게 수정 2

* Refactor: 일단 PickApplication이 돌아는 가게 수정 2에 같이 커밋 되었어야 할 파일

* Refactor: 없는 글 조회 시도 시 404 띄우는 코드

* Refactor: id들 Long으로 수정 외 다수

Feature: Comment 도메인 단일 댓글 조회 기능 구현
Refactor: Post 도메인에서 Comment 도메인 기능 분리
Refactor: 여러 변경점 반영...

* Refactor: main의 변경사항 반영 및 충돌 해결

* Refactor: Postman 돌려보며 Post 도메인 기능 정상 작동 확인

+++ Pick/src/main/java/com/nob/pick/post/command/domain/aggregate/vo/ResponseCommentVO.java
@@ -6,13 +6,13 @@ import lombok.Setter;
 @Getter
 @Setter
 public class ResponseCommentVO {
-	private int commentId;
+	private Long commentId;
 	private String commentIsAdopted;
 	private String commentUploadAt;
 	private String commentUpdateAt;
 	private String commentContent;
 	private int commentStatus;
-	private Integer commentRootCommentId;
-	private int commentMemberId;
+	private Long commentRootCommentId;
+	private Long commentMemberId;
 	private String commentMemberNickname;
 }
