+++ Pick/src/main/java/com/nob/pick/post/command/domain/aggregate/vo/ResponsePostListVO.java
@@ -6,12 +6,12 @@ import lombok.Setter;
 @Getter
 @Setter
 public class ResponsePostListVO {
-	private int id;
+	private Long id;
 	private String title;
 	private int category;
 	private String uploadAt;
 	private String updateAt;
 	private int status;
-	private int memberId;
+	private Long memberId;
 	private String memberNickname;
 }
